+++ ArsonistMod/SkillStates/ZeroPointBlast/ZeroPointBlastStart.cs
@@ -81,28 +81,15 @@ namespace ArsonistMod.SkillStates.ZeroPointBlast
             //Play Start/Whiff sound
             if (base.isAuthority)
             {
-                if (!Modules.Config.shouldHaveVoice.Value)
+                //Play Sound by default:
+                new PlaySoundNetworkRequest(characterBody.netId, 3585665340).Send(R2API.Networking.NetworkDestination.Clients);
+                //Soundbank has percentage chance set!
+                if (Modules.Config.shouldHaveVoice.Value && !characterBody.HasBuff(Modules.Buffs.masochismBuff)) 
                 {
-                    new PlaySoundNetworkRequest(characterBody.netId, 3585665340).Send(R2API.Networking.NetworkDestination.Clients);
+                    uint soundInt;
+                    soundInt = base.characterBody.skinIndex == Modules.Survivors.Arsonist.FirebugSkinIndex ? 2370694900 : 2004274107;
+                    new PlaySoundNetworkRequest(characterBody.netId, soundInt).Send(NetworkDestination.Clients);
                 }
-                else
-                {
-                    uint soundStr;
-                    
-                    //Firebug SFX check
-                    if (base.characterBody.skinIndex == Modules.Survivors.Arsonist.FirebugSkinIndex)
-                    {
-                        soundStr = characterBody.HasBuff(Modules.Buffs.masochismBuff) ? 3585665340 : 3167811629; //Nonlaugh : laugh
-                    }
-                    else 
-                    {
-                        //Determine if they have a buff and play a non-laughing version if so.
-                        soundStr = characterBody.HasBuff(Modules.Buffs.masochismBuff) ? 3585665340 : 289708206; //Nonlaugh : laugh
-                    }
-
-                    new PlaySoundNetworkRequest(characterBody.netId, soundStr).Send(R2API.Networking.NetworkDestination.Clients);
-                }
-                
             }
 
             energySystem = characterBody.gameObject.GetComponent<EnergySystem>();